requirements.txt
streamlit
google-generativeai
PyPDF2
python-docx
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
beautifulsoup4
requests
fastapi
uvicorn
python-multipart
